# redis client: default to use explicit pipelining