--- a/src/redis/driver.go
+++ b/src/redis/driver.go
@@ -1,5 +1,7 @@
 package redis
 
+import "github.com/mediocregopher/radix/v3"
+
 // Errors that may be raised during config parsing.
 type RedisError string
 
@@ -17,6 +19,22 @@ type Client interface {
 	// @param args supplies the additional arguments.
 	DoCmd(rcv interface{}, cmd, key string, args ...interface{}) error
 
+	// Pipe append a command onto the pipeline queue.
+	//
+	// @param pipeline supplies the queue for pending commands.
+	// @param rcv supplies receiver for the result.
+	// @param cmd supplies the command to append.
+	// @param key supplies the key to append.
+	// @param args supplies the additional arguments.
+	PipeAppend(pipeline Pipeline, rcv interface{}, cmd, key string, args ...interface{}) Pipeline
+
+	// PipeDo writes multiple commands to a Conn in
+	// a single write, then reads their responses in a single read. This reduces
+	// network delay into a single round-trip.
+	//
+	// @param pipeline supplies the queue for pending commands.
+	PipeDo(pipeline Pipeline) error
+
 	// Once Close() is called all future method calls on the Client will return
 	// an error
 	Close() error
@@ -24,3 +42,5 @@ type Client interface {
 	// NumActiveConns return number of active connections, used in testing.
 	NumActiveConns() int
 }
+
+type Pipeline []radix.CmdAction