--- a/test/mocks/redis/redis.go
+++ b/test/mocks/redis/redis.go
@@ -66,6 +66,20 @@ func (mr *MockClientMockRecorder) DoCmd(arg0, arg1, arg2 interface{}, arg3 ...in
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DoCmd", reflect.TypeOf((*MockClient)(nil).DoCmd), varargs...)
 }
 
+// ImplicitPipeliningEnabled mocks base method
+func (m *MockClient) ImplicitPipeliningEnabled() bool {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "ImplicitPipeliningEnabled")
+	ret0, _ := ret[0].(bool)
+	return ret0
+}
+
+// ImplicitPipeliningEnabled indicates an expected call of ImplicitPipeliningEnabled
+func (mr *MockClientMockRecorder) ImplicitPipeliningEnabled() *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ImplicitPipeliningEnabled", reflect.TypeOf((*MockClient)(nil).ImplicitPipeliningEnabled))
+}
+
 // NumActiveConns mocks base method
 func (m *MockClient) NumActiveConns() int {
 	m.ctrl.T.Helper()