Fix https://github.com/envoyproxy/ratelimit/issues/164

See conversations in https://github.com/envoyproxy/ratelimit/issues/152#issuecomment-668212174

Before 137 , explicit pipelining was used. After 137, we change to implicit pipelining to improve throughput (sacrifice some latency). This can cause problems for users who are concerned about latency. This pr make an optimization for this. If user turn off implicit pipelining by configuring pipeline_window=0, pipeline_limit=0, the redis client will use explicit pipelining.

Here is benchmark stats for `DoLimit` , for pipeline_window=0, pipeline_limit=0 case

Comparing to 1657473ad4a344de5428ec98b494d5112ef242ff
```
benchstat /tmp/ratelimit-before-1657473a.stat /tmp/ratelimit-opt.stat
name                           old time/op    new time/op    delta
ParallelDoLimit/no_pipeline-4    12.9µs ± 0%     9.1µs ± 0%   ~     (p=1.000 n=1+1)

name                           old alloc/op   new alloc/op   delta
ParallelDoLimit/no_pipeline-4      624B ± 0%      408B ± 0%   ~     (p=1.000 n=1+1)

name                           old allocs/op  new allocs/op  delta
ParallelDoLimit/no_pipeline-4      29.0 ± 0%      20.0 ± 0%   ~     (p=1.000 n=1+1)
```

Comparing to master

```
benchstat /tmp/ratelimit-master.stat /tmp/ratelimit-opt.stat 
name                           old time/op    new time/op    delta
ParallelDoLimit/no_pipeline-4    16.1µs ± 0%     9.1µs ± 0%   ~     (p=1.000 n=1+1)

name                           old alloc/op   new alloc/op   delta
ParallelDoLimit/no_pipeline-4      302B ± 0%      408B ± 0%   ~     (p=1.000 n=1+1)

name                           old allocs/op  new allocs/op  delta
ParallelDoLimit/no_pipeline-4      16.0 ± 0%      20.0 ± 0%   ~     (p=1.000 n=1+1)
```